pyproject.toml
[tool.poetry]
name = "doc_manager"
version = "0.1.0"
description = ""
authors = ["Kirill Andriychuk <kirill@key2market.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pypdf = "^5.1.0"
langchain = "^0.3.10"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.11"
langchain-community = "^0.3.14"
weaviate-client = "^4.10.3"
pytest = "^8.3.4"
langchain-ollama = "^0.2.2"
langchain-weaviate = "^0.0.3"
pytesseract = "^0.3.13"
pillow = "^11.1.0"
langchain-aws = "^0.2.11"
boto3 = "^1.36.3"
langdetect = "^1.0.9"
pdf2image = "^1.17.0"
pytz = "^2024.2"
opencv-python = "^4.11.0.86"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
